#74 made SmvKeys an object
