+++ src/main/scala/org/tresamigos/smv/SmvKeys.scala
@@ -14,10 +14,11 @@
 
 package org.tresamigos.smv
 
-import org.apache.spark.sql.{Column, DataFrame}
 import org.apache.spark.sql.types.{Metadata, MetadataBuilder, StructField}
 
-private[smv] class SmvKeys {
+/** Utility for using the metadata part of StructField to store Smv keys 
+ */
+private[smv] object SmvKeys {
   val SmvLabel = "smvLabel"
   val SmvDesc  = "smvDesc"
 
@@ -54,108 +55,5 @@ private[smv] class SmvKeys {
   def removeDescFromMeta(m: Metadata): Metadata = {
     val builder = new MetadataBuilder().withMetadata(m)
     builder.putString(SmvDesc, "").build
-  }
-}
-
-private[smv] class SchemaMetaOps(df: DataFrame) extends SmvKeys {
-
-  /**
-   * Adds labels to the specified columns.
-   *
-   * A column may have multiple labels.  Adding the same label twice
-   * to a column has the same effect as adding that label once.
-   *
-   * For multiple colNames, the same set of labels will be added to all of them.
-   * When colNames is empty, the set of labels will be added to all columns of the df.
-   * labels parameters must be non-empty.
-   */
-  def addLabel(colNames: Seq[String], labels: Seq[String]): DataFrame = {
-    require(!labels.isEmpty)
-    val allCol = colNames.isEmpty
-
-    val columns = df.schema.fields map { f =>
-      val c = f.name
-
-      // if new label should be added to this column. Add to all columns, if colNames is empty
-      if (allCol || colNames.contains(c)) {
-        val meta = addLabelsToMeta(f.metadata, labels)
-        df(c).as(c, meta)
-      } else df(f.name)
-    }
-
-    df.select(columns: _*)
-  }
-
-  def addDesc(colDescs: Seq[(String, String)]): DataFrame = {
-    require(!colDescs.isEmpty)
-    val colMap = colDescs.toMap
-
-    val columns = df.schema.fields map { f =>
-      val c = f.name
-      if (colMap.contains(c)) {
-        val meta = addDescToMeta(f.metadata, colMap.getOrElse(c, ""))
-        df(c).as(c, meta)
-      } else df(c)
-    }
-
-    df.select(columns: _*)
-  }
-
-  def getLabel(colName: String): Seq[String] = {
-    getMetaLabels(df.schema.apply(colName).metadata)
-  }
-
-  def getDesc(colName: String): String = {
-    getMetaDesc(df.schema.apply(colName).metadata)
-  }
-
-  def removeLabel(colNames: Seq[String], labels: Seq[String]): DataFrame = {
-    val allCol    = colNames.isEmpty
-
-    val columns = df.schema.fields map { f =>
-      val c = f.name
-      if (allCol || colNames.contains(c)) {
-        val meta = removeLabelsFromMeta(f.metadata, labels)
-        df(c) as (c, meta)
-      } else df(c)
-    }
-    df.select(columns: _*)
-  }
-
-  def removeDesc(colNames: Seq[String]): DataFrame = {
-    val allCol = colNames.isEmpty
-
-    val columns = df.schema.fields map { f =>
-      val c = f.name
-      if (allCol || colNames.contains(c)) {
-        val meta    = removeDescFromMeta(f.metadata)
-        df(c) as (c, meta)
-      } else df(c)
-    }
-    df.select(columns: _*)
-  }
-
-  def colWithLabel(labels: Seq[String]): Seq[String] = {
-    val filterFn: Metadata => Boolean = { meta =>
-      if (labels.isEmpty)
-        !meta.contains(SmvLabel) || meta.getStringArray(SmvLabel).isEmpty
-      else
-        meta.contains(SmvLabel) && labels.toSet.subsetOf(meta.getStringArray(SmvLabel).toSet)
-    }
-
-    val ret = for {
-      f <- df.schema.fields if (filterFn(f.metadata))
-    } yield f.name
-
-    require(
-      !ret.isEmpty,
-      if (labels.isEmpty)
-        s"""there are no unlabeled columns in the data frame [${df.columns.mkString(",")}]"""
-      else
-        s"""there are no columns labeled with ${labels} in the data frame [${df.columns.mkString(
-          ",")}]"""
-    )
-
-    ret
   }
 }
